// src/Model/ViewFeeStructure/FeeStructureItem.tsx
export interface FeeStructureItem {
  id: string | null;
  student_code: string | null;
  form_no: string | null;
  month_id: string | null;
  fine: string | null;
  admission_fee: string | null;
  development_fee: string | null;
  exam_fee: string | null;
  festival_celebration_fee: string | null;
  games_sports_fee: string | null;
  audio_visual_lab_fee: string | null;
  library_fee: string | null;
  electricity_maintenance_fee: string | null;
  computer_fee: string | null;
  session_charges: string | null;
  security_deposite: string | null;
  tuition_fee: string | null;
  bus_services: string | null;
  bus_fee_fine: string | null;
  payment_amount: string | null;
  academic_payment_amt: string | null;
  bus_payment_amt: string | null;
  total_stationary_fee: string | null;
  ad_payment_status: string | null;
  bus_payment_status: string | null;
  payment_due_date: string | null;
  created_date: string | null;
  bus_payment_date: string | null;
  ad_payment_mode: string | null;
  bus_payment_mode: string | null;
  cheque_number: string | null;
  pos_bank_name: string | null;
  pos_reference_number: string | null;
  transaction_no: string | null;
  payee_name: string | null;
  t_user_id: string | null;
  added_by: string | null;
  bus_cheque_number: string | null;
  bus_pos_bank_name: string | null;
  bus_pos_reference_number: string | null;
  bus_payee_name: string | null;
  bus_t_user_id: string | null;
  bus_added_by: string | null;
  cons_admission_fee: string | null;
  cons_development_fee: string | null;
  cons_exam_fee: string | null;
  cons_festival_celebration_fee: string | null;
  cons_games_sports_fee: string | null;
  cons_audio_visual_lab_fee: string | null;
  cons_library_fee: string | null;
  cons_electricity_maintenance_fee: string | null;
  cons_computer_fee: string | null;
  cons_session_charges: string | null;
  cons_security_deposite: string | null;
  cons_tuition_fee: string | null;
  cons_bus_services: string | null;
  remarks: string | null;
  adv_bal_used: string | null;
  session_year_id: string | null;
  allow_bus_report: string | null;
  allow_payment_report: string | null;
  cons_pdf: string | null;
  fee_id: string | null;
  class_name: string | null;
  first_name: string | null;
  roll_num: string | null;
  father_name: string | null;
  mother_name: string | null;
  month_name: string;
  total_amount_show: string;
}